slightly smarter raycasting, prioritising found planes, then estimated planes
